# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "F1 Reaction Time API",
    "version": "1.0.0",
    "description": "API documentation for F1 Reaction Time"
  },
  "paths": {
    "/submit-reaction-time": {
      "post": {
        "summary": "Submit reaction time",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reactionTime": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A successful response : reaction time submitted"
          }
        }
      }
    },
    "/get-reaction-times/{userId}": {
      "get": {
        "summary": "Get reaction times for a user",
        "description": "Fetches the reaction times for a specific user, sorted by the most recent.",
        "tags": ["Reaction Times"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "ID of the user whose reaction times are being fetched",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of reaction times",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "user_id": {
                        "type": "string",
                        "description": "ID of the user"
                      },
                      "time": {
                        "type": "number",
                        "description": "The recorded reaction time"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Timestamp of when the reaction time was recorded"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}